Backend/requirements.txt
flask
mysql-connector-python
pandas
tqdm
click